+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.2.0
+
 [GD-2045](https://jira.gutefrage.net/browse/GD-2045). Add liveramp TCF 2 module
 [GD-2044](https://jira.gutefrage.net/browse/GD-2044). Add gdprEnforcement types to prebid.ts